LayoutTests/http/tests/websocket/tests/error-detect.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../js-test-resources/js-test-style.css">
<script src="../../../js-test-resources/js-test-pre.js"></script>
<script src="../../../js-test-resources/js-test-post-function.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure WebSocket correctly fire error event for unknown frame type.");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

var errorCount = 0;

function finish() {
    shouldBe("errorCount", "255");
    isSuccessfullyParsed();
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/unknown-frame-type");
ws.onopen = function () {
    debug("WebSocket is open");
};
ws.onmessage = function (evt) {
    debug("received:" + evt.data);
};
ws.onerror = function () {
    errorCount += 1;
};
ws.onclose = function () {
    debug("WebSocket is closed");
    finish();
};

var successfullyParsed = true;
</script>
</body>
</html>
