safari-534.53-branch/LayoutTests/http/tests/websocket/tests/error-detect-expected.txt

Make sure WebSocket correctly fire error event for unknown frame type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket is open
received:done
WebSocket is closed
PASS errorCount is 255
PASS successfullyParsed is true

TEST COMPLETE
